# backend/grip/federation/contract/opdrachtverkeer/v1/openapi.yaml
openapi: 3.1.0
info:
  title: grip-opdrachtverkeer
  version: 1.0.0
  summary: Het verkeer tussen grip-instanties over aanvragen, offertes, akkoord, inzage, vacatures en doorgifte.
  description: |
    Elke grip-instantie biedt deze dienst aan achter de eigen FSC-inway, onder de dienstnaam
    `grip-opdrachtverkeer`. Een instantie kan opdrachtgever en opdrachtnemer zijn. Een opdracht is
    lokaal volledig; dit verkeer is een laag erbovenop.

    **Identiteit.** Opdrachten, offertes en vacatures hebben een URI van de instantie die ze
    aanmaakt, bijvoorbeeld `{basis}/id/opdracht/{uuid}`. De `{opdrachtId}`, `{offerteId}` en
    `{vacatureId}` in de paden zijn de `{uuid}` uit die URI. Organisaties worden aangeduid met hun
    TOOI-URI.

    **Aanroeper.** De aanroeper is een FSC-peer. De ontvanger leidt de relatie af uit het peer-id:
    opdrachtgever of opdrachtnemer van die opdracht, moederinstantie, of corpus. Elke operatie
    noemt in `x-grip-caller` welke relatie vereist is. Een opdracht waar de aanroeper geen relatie
    mee heeft geeft 404.

    **Gegevensklassen.** Elke operatie noemt in `x-grip-data-class` de klassen die het bericht of
    antwoord bevat: A opdracht-basis, B opdracht-financieel, C bemensing, D persoon-tarief,
    E persoon-kosten, F persoon-KPI. De klassen D, E en F gaan nooit over de grens van een
    instantie. Naar een opdrachtgever gaan geen namen van medewerkers; klasse C met namen gaat
    alleen naar de moederinstantie.

...

...

...

...
  contact:
...
servers:
...
paths:
  /opdrachtaanvragen:
    post:
      operationId: sendAssignmentRequest
      tags: [Aanvraag en offerte]
      summary: Aanvraag sturen
      description: De opdrachtgever vraagt de opdrachtnemer om een offerte. De context is een lijst node-URIs en mag leeg zijn. De opdrachtgever geeft de URI van de opdracht uit.
      x-grip-caller: client
      x-grip-data-class: [A]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../../schemas/opdrachtaanvraag.schema.json" }
      responses:
        "201": { $ref: "#/components/responses/Received" }
        "200": { $ref: "#/components/responses/Duplicate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /offertes:
    post:
      operationId: sendQuote
      tags: [Aanvraag en offerte]
      summary: Offerte sturen
      description: |
        De opdrachtnemer stuurt een uitgegeven offerte: een bevroren momentopname met de SHA-256 over
        de canonieke JSON (RFC 8785) van `momentopname`. De ontvanger rekent de hash na en weigert bij
        een verschil met 400.

        Een offerte mag zonder voorafgaande aanvraag komen (`aanvraag_id` is dan null); de
        opdrachtnemer geeft in dat geval de URI van de opdracht uit. Een gewijzigde offerte is een
        nieuwe offerte met een nieuw `id`.
      x-grip-caller: contractor
      x-grip-data-class: [A, B]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../../schemas/offerte.schema.json" }
      responses:
        "201": { $ref: "#/components/responses/Received" }
        "200": { $ref: "#/components/responses/Duplicate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /offertes/{offerteId}/akkoorden:
    post:
      operationId: sendAcceptance
      tags: [Aanvraag en offerte]
      summary: Akkoord sturen
      description: |
        De opdrachtgever geeft akkoord op een offerte. Het bericht legt vast wie tekende, namens
        welke organisatie, wanneer, op welke versie (`offerte_hash`) en in welke vorm.

        Bij `vorm` `eigen_instantie` ondertekent de instantie van de opdrachtgever het bericht met een
        JWS. De ontvanger controleert de handtekening met de sleutels uit `GET /jwks` van de
        afzender, en controleert dat de payload gelijk is aan de velden van het bericht. De vormen
        `tekenlink` en `geuploade_pdf` ontstaan bij de opdrachtnemer zelf; ze staan in het schema
        zodat een akkoord in elke vorm kan worden doorgegeven in een keten.

        Een `offerte_hash` die niet hoort bij de uitgegeven offerte geeft 409, net als een offerte
        waarop al een akkoord of afwijzing met een ander `id` is ontvangen.
      x-grip-caller: client
      x-grip-data-class: [A]
      parameters:
        - { $ref: "#/components/parameters/offerteId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../../schemas/akkoord.schema.json" }
      responses:
        "201": { $ref: "#/components/responses/Received" }
        "200": { $ref: "#/components/responses/Duplicate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /offertes/{offerteId}/afwijzingen:
    post:
      operationId: sendRejection
      tags: [Aanvraag en offerte]
      summary: Afwijzing sturen
      description: De opdrachtgever wijst een offerte af. Dezelfde regels voor `offerte_hash` en 409 als bij een akkoord.
      x-grip-caller: client
      x-grip-data-class: [A]
      parameters:
        - { $ref: "#/components/parameters/offerteId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../../schemas/afwijzing.schema.json" }
      responses:
        "201": { $ref: "#/components/responses/Received" }
        "200": { $ref: "#/components/responses/Duplicate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /opdrachten:
    get:
      operationId: listAssignmentsByNode
      tags: [Corpus]
      summary: Opdrachten die naar een node verwijzen
      description: Geeft de opdrachten waarvan `context_uris` deze node-URI bevat. Bedoeld voor een corpus-systeem dat opdrachten alleen-lezen bij een node toont. Alleen kerngegevens; geen bedragen en geen namen.
      x-grip-caller: corpus
      x-grip-data-class: [A]
      parameters:
        - name: nodeUri
          in: query
          required: true
          description: Volledige URI van de node.
          schema: { type: string, format: uri }
        - { $ref: "#/components/parameters/page" }
        - { $ref: "#/components/parameters/pageSize" }
      responses:
        "200":
          description: Opdrachten bij deze node.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/opdracht-pagina.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /opdrachten/{opdrachtId}:
    get:
      operationId: getAssignment
      tags: [Inzage]
      summary: Kerngegevens van een opdracht
      x-grip-caller: client
      x-grip-data-class: [A]
      parameters:
        - { $ref: "#/components/parameters/opdrachtId" }
      responses:
        "200":
          description: De opdracht.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/opdracht.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /opdrachten/{opdrachtId}/voortgang:
    get:
      operationId: getProgress
      tags: [Inzage]
      summary: Status en voortgang
      description: Standaard opvraagbaar door de opdrachtgever. Geen bedragen en geen namen.
      x-grip-caller: client
      x-grip-data-class: [A]
      parameters:
        - { $ref: "#/components/parameters/opdrachtId" }
      responses:
        "200":
          description: De voortgang.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/voortgang.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /opdrachten/{opdrachtId}/uitputting:
    get:
      operationId: getBudgetUsage
      tags: [Inzage]
      summary: Uitputting
      description: Begroot, uitputting en beschikbaar. Alleen op verzoek; 403 als de opdrachtnemer dit voor deze opdrachtgever niet heeft toegestaan.
      x-grip-caller: client
      x-grip-data-class: [B]
      x-grip-on-request: true
      parameters:
        - { $ref: "#/components/parameters/opdrachtId" }
        - name: jaar
          in: query
          description: Kalenderjaar. Zonder deze parameter de hele looptijd.
          schema: { type: integer, minimum: 2000 }
      responses:
        "200":
          description: De uitputting.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/uitputting.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /opdrachten/{opdrachtId}/factuurgegevens:
    get:
      operationId: getBillingData
      tags: [Inzage]
      summary: Factuurgegevens van een maand
      description: Factuurgegevens van een afgesloten maand. Alleen op verzoek; 403 als dat niet is toegestaan, 404 als de maand nog niet is afgesloten.
      x-grip-caller: client
      x-grip-data-class: [B]
      x-grip-on-request: true
      parameters:
        - { $ref: "#/components/parameters/opdrachtId" }
        - { $ref: "#/components/parameters/maand" }
      responses:
        "200":
          description: De factuurgegevens.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/factuurgegevens.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /opdrachten/{opdrachtId}/eindrapport:
    put:
      operationId: sendFinalReport
      tags: [Eindrapport]
      summary: Eindrapport sturen
      description: De opdrachtnemer stuurt het eindrapport naar de opdrachtgever. De aanroep gaat naar de instantie van de opdrachtgever. Een herzien rapport heeft een nieuw `id` en vervangt het vorige.
      x-grip-caller: contractor
      x-grip-data-class: [A, B]
      parameters:
        - { $ref: "#/components/parameters/opdrachtId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../../schemas/eindrapport.schema.json" }
      responses:
        "201": { $ref: "#/components/responses/Received" }
        "200": { $ref: "#/components/responses/Duplicate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /vacatures:
    post:
      operationId: sendVacancy
      tags: [Vacatures]
      summary: Vacature uitzetten
      description: Een instantie zet een open rol uit bij een andere instantie waarmee ze een contract heeft. Een wijziging van de stand (ingevuld, ingetrokken) is hetzelfde bericht met een nieuw `id` en dezelfde `uri`.
      x-grip-caller: peer
      x-grip-data-class: [A]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../../schemas/vacature.schema.json" }
      responses:
        "201": { $ref: "#/components/responses/Received" }
        "200": { $ref: "#/components/responses/Duplicate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /vacatures/{vacatureId}/aanbiedingen:
    post:
      operationId: sendVacancyOffer
      tags: [Vacatures]
      summary: Kandidaat aanbieden
      description: Een instantie biedt een kandidaat aan voor een ontvangen vacature. De aanroep gaat naar de instantie die de vacature uitzette. Het aanbod bevat geen naam.
      x-grip-caller: peer
      x-grip-data-class: [C]
      parameters:
        - name: vacatureId
          in: path
          required: true
          description: De uuid uit de vacature-URI.
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../../schemas/aanbod.schema.json" }
      responses:
        "201": { $ref: "#/components/responses/Received" }
        "200": { $ref: "#/components/responses/Duplicate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /doorgifte/opdrachten:
    get:
      operationId: getHandoverAssignments
      tags: [Doorgifte]
      summary: Opdrachten met status en bedragen
      x-grip-caller: parent
      x-grip-data-class: [A, B]
      responses:
        "200":
          description: Alle opdrachten van dit onderdeel.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/doorgifte-opdrachten.schema.json" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /doorgifte/factuurgegevens:
    get:
      operationId: getHandoverBillingData
      tags: [Doorgifte]
      summary: Factuurgegevens van een maand, alle opdrachten
      x-grip-caller: parent
      x-grip-data-class: [B]
      parameters:
        - { $ref: "#/components/parameters/maand" }
      responses:
        "200":
          description: Factuurgegevens per opdracht.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/doorgifte-factuurgegevens.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /doorgifte/bemensing:
    get:
      operationId: getHandoverStaffing
      tags: [Doorgifte]
      summary: Bezetting met namen
      description: Inzet per persoon per opdracht in een maand. Alleen voor de moederinstantie; dit is de enige operatie die namen van medewerkers bevat.
      x-grip-caller: parent
      x-grip-data-class: [C]
      parameters:
        - { $ref: "#/components/parameters/maand" }
      responses:
        "200":
          description: De bezetting.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/doorgifte-bemensing.schema.json" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /doorgifte/capaciteit:
    get:
      operationId: getHandoverCapacity
      tags: [Doorgifte]
      summary: Capaciteit in totalen
      x-grip-caller: parent
      x-grip-data-class: [C]
      responses:
        "200":
          description: Ingezette, beschikbare en gezochte capaciteit.
          headers:
            API-Version: { $ref: "#/components/headers/API-Version" }
          content:
            application/json:
              schema: { $ref: "../../schemas/doorgifte-capaciteit.schema.json" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
...
  responses:
    Received:
      description: Het bericht is voor het eerst ontvangen en vastgelegd.
      headers:
        API-Version: { $ref: "#/components/headers/API-Version" }
      content:
        application/json:
          schema: { $ref: "../../schemas/ontvangstbevestiging.schema.json" }
...